dom/media/webaudio/test/test_decoderDelay.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test that decoder delay is handled</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<pre id="test">
<script class="testbody" type="text/javascript">

SimpleTest.waitForExplicitFinish();

var tests = [
  {
    name: "half-a-second-8000.mp3",
    duration: 0.5
  },
  {
    name: "half-a-second-48000.mp3",
    duration: 0.5
  },
  {
    name: "v3.mp3",
    duration: 3
  },
  {
    // Weird file that has the end padding larger than a packet size.
    name: "cropped_8000hz_8kbs_mono_lame3.100.abr.mp3",
    duration: 28.65675
  }
];

async function doit(t) {
  var count = 0;
  var context = new OfflineAudioContext(1, 128, 48000);
  tests.forEach(async testcase => {
    var response = await fetch(testcase.name);
    var buffer = await response.arrayBuffer();
    var decoded = await context.decodeAudioData(buffer);
    is(decoded.duration, testcase.duration, "The file has the correct duration.");
    if (++count == tests.length) {
      SimpleTest.finish();
    }
  });
}

doit();

</script>
</pre>
</body>
</html>
